Move cli_setup, connect_cli and host_cli to cli module
